+++ frontend/src/schedulegeneration/SGWorker.js
@@ -9,6 +9,8 @@ export function SGWorker() {
             return;
 
         let data = evt.data;
+        console.log("Data for this run is:");
+        console.log(data);
         let {classData, classTypesToIgnore, preferences} = data;
         // have to convert from JSON preferences to preference objects
         let {specificPref, globalPref} = initPreferences(preferences);
@@ -132,6 +134,9 @@ export function SGWorker() {
             startPref.setHours(startPref.getHours(), startPref.getMinutes(), 0);
             endPref.setHours(endPref.getHours(), endPref.getMinutes(), 0);
 
+            console.log(startPref);
+            console.log(endPref);
+
             let score = 0;
             // this shouldn't happen but if it does punish hardcore
             if (!subsection.timeInterval) {
@@ -166,6 +171,7 @@ export function SGWorker() {
 
         GlobalPref.prototype.evaluateDay = function (subsection) {
             let dayPref = this.globalPref.dayPref;
+
             if (!dayPref)
                 return 0;
