+++ frontend/src/utils/message/ConsoleMessageHandler.js
@@ -2,11 +2,17 @@
 
 export class ConsoleMessageHandler {
 
+    constructor() {
+        this.messageQueue = []
+    }
+
     showError(message) {
+        this.messageQueue.push(message);
         console.error(message);
     }
 
     showSuccess(message) {
+        this.messageQueue.push(message);
         console.log(message);
     }
 }
